+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 
+### Added
+- Added new syntax for `SetFee` psuedo transaction sent after the [XRPFees](https://xrpl.org/known-amendments.html#xrpfees) amendment. (Backwards compatible)
+- Update request models related to AMM
+
 ## [2.3.0] - 2023-08-24
 ### Added
 - Add AMM support [XLS-30](https://github.com/XRPLF/XRPL-Standards/discussions/78)